// src/shared/enums/role.enum.ts
export enum Role {
  ADMIN = 'ADMIN',
  CEO = 'CEO',
  MANAGER = 'MANAGER',
  ACCOUNTING = 'ACCOUNTING',
  DEVELOPER = 'DEVELOPER',
  SOCIAL_MEDIA = 'SOCIAL_MEDIA',
  CREATIVE = 'CREATIVE',
  MARKETING = 'MARKETING',
  PRODUCTION = 'PRODUCTION',
  ACCOUNT_MANAGER = 'ACCOUNT_MANAGER',
  HR = 'HR',
  CLIENT = 'CLIENT',
}